+++ british_food_generator/models.py
@@ -1,4 +1,21 @@
+from constrained_types import add_constraint, ConstrainedString
 from pydantic import BaseModel
+
+
+@add_constraint(lambda s: len(s) > 0, "There was an empty url")
+@add_constraint(lambda s: s.startswith("https://"), "The url should start with https")
+class ImgUrl(ConstrainedString):
+    pass
+
+
+@add_constraint(lambda s: len(s) > 0, "The description should not be empty")
+class FoodDescription(ConstrainedString):
+    pass
+
+
+@add_constraint(lambda s: len(s) > 0, "The name should not be empty")
+class FoodName(ConstrainedString):
+    pass
 
 
 class ClassicBritishDish(BaseModel):
@@ -6,9 +23,9 @@ class ClassicBritishDish(BaseModel):
     A tasty and definitely real part of British cuisine
     """
 
-    name: str
-    description: str
-    image: str
+    name: FoodName
+    description: FoodDescription
+    image: ImgUrl
 
     class Config:
         schema_extra = {
